add: config/chprompt_bashrc.txt
add: metadata/chbashrc.json

add: metadata/init.json

add: module/chbashrc.sh

add: module/init.sh

add: utils/lhome.sh

add: utils/perl/audit_plugin.pl

add: utils/perl/json_parser.pl

modified: .docs/install_and_uninstall.md

modified: .install/packages.txt

modified: bin/chprompt.sh

modified: chprompt.sh

modified: module/audit.sh

modified: module/current_prompt.sh

modified: module/helper.sh

modified: module/inject.sh

modified: uninstall.sh

modified: utils/cprompt.sh

remove: utils/python/audit_plugin.py

remove: utils/python/json_parser.py

+++ uninstall.sh
@@ -30,17 +30,23 @@ include : '(
 )'
 
 HOME="${HOME}"
+__RMBK__=false
+__RMC__=false
 
 while [[ ${#} -gt 0 ]]; do
     case "${1}" in
         "--home="*) export HOME="${1#*=}" ;;
+        "--remove-backup") export __RMBK__=true ;;
+        "--no-remove-code") export __RMC__=true ;;
     esac
     shift
 done
 
-install::getinstall \
-    "command rm -f ${opt}/chprompt_*.zip.bak" \
-    "Removing all backup..."
+if [[ "${__RMBK__}" == true ]]; then
+    install::getinstall \
+        "command rm -f ${opt}/chprompt_*.zip.bak" \
+        "Removing all backup..."
+fi
 
 install::getinstall \
     "command rm -rf ${opt}/chprompt" \
@@ -50,21 +56,23 @@ install::getinstall \
     "command rm -f ${bin}/chprompt" \
     "Removing: ${GG}${bin}/chprompt${N}"
 
-install::getinstall \
-    "
-        command cat ${HOME}/.bashrc | \
-            command grep -v 'chprompt' \
-            > ${tmp}/chprompt_uninstall
-    " \
-    "Filtering: ${GG}${HOME}/.bashrc ${GG}-> ${GG}${tmp}/chprompt_uninstall"
+if [[ "${__RMC__}" == false ]]; then
+    install::getinstall \
+        "
+            command cat ${HOME}/.bashrc | \
+                command grep -v 'chprompt' \
+                > ${tmp}/chprompt_uninstall
+        " \
+        "Filtering: ${GG}${HOME}/.bashrc ${GG}-> ${GG}${tmp}/chprompt_uninstall"
 
-install::getinstall \
-    "
-        command mv \
-            ${tmp}/chprompt_uninstall \
-            ${HOME}/.bashrc
-    " \
-    "Moving: ${GG}${tmp}/chprompt_uninstall ${DG}-> ${GG}${HOME}/.bashrc${N}"
+    install::getinstall \
+        "
+            command mv \
+                ${tmp}/chprompt_uninstall \
+                ${HOME}/.bashrc
+        " \
+        "Moving: ${GG}${tmp}/chprompt_uninstall ${DG}-> ${GG}${HOME}/.bashrc${N}"
+fi
 
 echo -e "${GG}[+] ${N}Chprompt removed"
